Add some extensions to pandocCompiler
Pandoc and markdown-unlit does not play too well.

markdown-unlit accept the following and ignores it as haskell code

```haskell ignore
main :: IO ()
main = putStrLn "Hello World!"
```

but Pandoc treats that as a Para block and **not** a CodeBlock.
I found no way to make it easier and only workaround seemed to be to
instead transform the above to this

```{.haskell .ignore}
main :: IO ()
main = putStrLn "Hello World!"
```

and that parses fine in both tools. markdown-unlit ignores it and
pandoc parses that as an actual CodeBlock with the classes "haskell ignore".